@@ -4,18 +4,17 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0<3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^0.1.2
-  animated_text_kit: ^1.3.0
-  firebase_core: ^0.3.4
-  firebase_auth: ^0.8.4+4
-  cloud_firestore: ^0.9.13+1
-  modal_progress_hud: ^0.1.3
+  cupertino_icons: ^1.0.5
+  animated_text_kit: ^4.2.1
+  firebase_core: ^1.12.0
+  firebase_auth: ^3.3.8
+  cloud_firestore: ^3.1.9
 
 dev_dependencies:
   flutter_test:
