@@ -3,7 +3,6 @@ import 'package:flash_chat/screens/chat_screen.dart';
 import 'package:flash_chat/wdigets/RoundedButton.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:modal_progress_hud/modal_progress_hud.dart';
 
 class LoginScreen extends StatefulWidget {
   static String id = 'login_screen';
@@ -15,83 +14,73 @@ class _LoginScreenState extends State<LoginScreen> {
 
   final FirebaseAuth _auth = FirebaseAuth.instance;
 
-  String email;
-  String password;
+  late String email;
+  late String password;
 
-  bool showSpinner = false;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.white,
-      body: ModalProgressHUD(
-        inAsyncCall: showSpinner,
-        child: Padding(
-          padding: EdgeInsets.symmetric(horizontal: 24.0),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: <Widget>[
-              Hero(
-                tag: 'logo',
-                child: Container(
-                  height: 200.0,
-                  child: Image.asset('images/logo.png'),
-                ),
+      body: Padding(
+        padding: EdgeInsets.symmetric(horizontal: 24.0),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: <Widget>[
+            Hero(
+              tag: 'logo',
+              child: Container(
+                height: 200.0,
+                child: Image.asset('images/logo.png'),
               ),
-              SizedBox(
-                height: 48.0,
+            ),
+            SizedBox(
+              height: 48.0,
+            ),
+            TextField(
+              textAlign: TextAlign.center,
+              style: TextStyle(
+                color: Colors.black
               ),
-              TextField(
-                textAlign: TextAlign.center,
-                style: TextStyle(
-                  color: Colors.black
-                ),
-                onChanged: (value) {
-                  email = value;
-                },
-                decoration: kTextFieldDecoration.copyWith(hintText: 'email'),
+              onChanged: (value) {
+                email = value;
+              },
+              decoration: kTextFieldDecoration.copyWith(hintText: 'email'),
+            ),
+            SizedBox(
+              height: 8.0,
+            ),
+            TextField(
+              obscureText: true,
+              textAlign: TextAlign.center,
+              style: TextStyle(
+                color: Colors.black
               ),
-              SizedBox(
-                height: 8.0,
+              onChanged: (value) {
+                password = value;
+              },
+              decoration: kTextFieldDecoration.copyWith(hintText: 'password'),
+            ),
+            SizedBox(
+              height: 24.0,
+            ),
+            RoundedButton(
+              text: 'Log in',
+              color: Colors.lightBlueAccent,
+              onpressed: ()async{
+                try{
+                  final user = await _auth.signInWithEmailAndPassword(email: email, password: password);
+                  if(user != null){
+                    Navigator.pushNamed(context, ChatScreen.id);
+                }
+                }
+                  catch(e){
+                    print('Error occured while signing in');
+                }
+              },              
               ),
-              TextField(
-                obscureText: true,
-                textAlign: TextAlign.center,
-                style: TextStyle(
-                  color: Colors.black
-                ),
-                onChanged: (value) {
-                  password = value;
-                },
-                decoration: kTextFieldDecoration.copyWith(hintText: 'password'),
-              ),
-              SizedBox(
-                height: 24.0,
-              ),
-              RoundedButton(
-                text: 'Log in',
-                color: Colors.lightBlueAccent,
-                onpressed: ()async{
-                  setState(() {
-                    showSpinner = true;
-                  });
-                  try{
-                    final user = await _auth.signInWithEmailAndPassword(email: email, password: password);
-                    if(user != null){
-                      Navigator.pushNamed(context, ChatScreen.id);
-                  }
-                  setState(() {
-                    showSpinner = false;                    
-                  });
-                  }
-                    catch(e){
-                      print('Error occured while signing in');
-                  }
-                },              
-                ),
-            ],
-          ),
+          ],
         ),
       ),
     );
