@@ -3,7 +3,6 @@ import 'package:flash_chat/screens/chat_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flash_chat/wdigets/RoundedButton.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:modal_progress_hud/modal_progress_hud.dart';
 
 class RegistrationScreen extends StatefulWidget {
   static String id = 'registration_screen';
@@ -13,84 +12,74 @@ class RegistrationScreen extends StatefulWidget {
 
 class _RegistrationScreenState extends State<RegistrationScreen> {
 
-  String email;
-  String password;
-  bool showspinner = false;
+  late String email;
+  late String password;
   final FirebaseAuth _auth = FirebaseAuth.instance;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.white,
-      body: ModalProgressHUD(
-        inAsyncCall: showspinner,
-        child: Padding(
-          padding: EdgeInsets.symmetric(horizontal: 24.0),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: <Widget>[
-              Hero(
-                tag: 'logo',
-                child: Container(
-                  height: 200.0,
-                  child: Image.asset('images/logo.png'),
-                ),
+      body: Padding(
+        padding: EdgeInsets.symmetric(horizontal: 24.0),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: <Widget>[
+            Hero(
+              tag: 'logo',
+              child: Container(
+                height: 200.0,
+                child: Image.asset('images/logo.png'),
               ),
-              SizedBox(
-                height: 48.0,
+            ),
+            SizedBox(
+              height: 48.0,
+            ),
+            TextField(
+              textAlign: TextAlign.center,
+              keyboardType: TextInputType.emailAddress,
+              style: TextStyle(
+                color: Colors.black
               ),
-              TextField(
-                textAlign: TextAlign.center,
-                keyboardType: TextInputType.emailAddress,
-                style: TextStyle(
-                  color: Colors.black
-                ),
-                onChanged: (value) {
-                  email = value;
-                },
-                decoration: kTextFieldDecoration.copyWith(hintText: 'email')
+              onChanged: (value) {
+                email = value;
+              },
+              decoration: kTextFieldDecoration.copyWith(hintText: 'email')
+            ),
+            SizedBox(
+              height: 8.0,
+            ),
+            TextField(
+              textAlign: TextAlign.center,
+              style: TextStyle(
+                color: Colors.black
               ),
-              SizedBox(
-                height: 8.0,
+              obscureText: true,
+              onChanged: (value) {
+                password = value;
+              },
+              decoration: kTextFieldDecoration.copyWith(hintText: 'password'),
+            ),
+            SizedBox(
+              height: 24.0,
+            ),
+            RoundedButton(
+              color: Colors.blueAccent, 
+              onpressed: ()async{
+                try{
+                final newUser = await _auth.createUserWithEmailAndPassword(email: email, password: password);
+                if(newUser != null){
+                  Navigator.pushNamed(context, ChatScreen.id);
+                }
+                }
+                catch(e){
+                  print('An error Occured');
+                }
+              }, 
+              text: 'Register'
               ),
-              TextField(
-                textAlign: TextAlign.center,
-                style: TextStyle(
-                  color: Colors.black
-                ),
-                obscureText: true,
-                onChanged: (value) {
-                  password = value;
-                },
-                decoration: kTextFieldDecoration.copyWith(hintText: 'password'),
-              ),
-              SizedBox(
-                height: 24.0,
-              ),
-              RoundedButton(
-                color: Colors.blueAccent, 
-                onpressed: ()async{
-                  setState(() {
-                    showspinner = true;
-                  });
-                  try{
-                  final newUser = await _auth.createUserWithEmailAndPassword(email: email, password: password);
-                  if(newUser != null){
-                    Navigator.pushNamed(context, ChatScreen.id);
-                  }
-                  setState(() {
-                    showspinner = false;
-                  });
-                  }
-                  catch(e){
-                    print('An error Occured');
-                  }
-                }, 
-                text: 'Register'
-                ),
-            ],
-          ),
+          ],
         ),
       ),
     );
